change func IdNum's return of valueobject Id, uint64 to uint

+++ common/domain/primitive/primitive.go
@@ -11,7 +11,7 @@ const timeFormatYearToSecond = "2006-01-02 15:04:05"
 // id
 type Id interface {
 	Id() string
-	IdNum() uint64
+	IdNum() uint
 }
 
 type id string
@@ -26,9 +26,9 @@ func (r id) Id() string {
 
 // IdNum return an uint type ID.
 // return 0 if it is unable to convert
-func (r id) IdNum() uint64 {
+func (r id) IdNum() uint {
 	if num, err := strconv.ParseUint(r.Id(), 10, 64); err == nil {
-		return num
+		return uint(num)
 	}
 
 	return 0
